web-application/requirements.txt
boto3==1.21.10
Django==4.0.3
django-storages==1.12.3
djangorestframework==3.13.1